Extract DTLS to dtls.c
